fix nodenext typings
