+++ types/index.d.ts
@@ -4,122 +4,118 @@ import {
   FastifyPluginCallback,
   FastifyRequest,
   preHandlerAsyncHookHandler,
-  RawRequestDefaultExpression,
-  RawServerBase,
-  RawServerDefault,
-  RequestGenericInterface,
   RouteOptions
 } from 'fastify';
 
 declare module 'fastify' {
-  interface FastifyRequestInterface<
-    RawServer extends RawServerBase = RawServerDefault,
-    RawRequest extends RawRequestDefaultExpression<RawServer> = RawRequestDefaultExpression<RawServer>,
-    RequestGeneric extends RequestGenericInterface = RequestGenericInterface
-  > {
-    ip: string | number;
-  }
   interface FastifyInstance {
-    rateLimit: (options?: RateLimitOptions) => preHandlerAsyncHookHandler;
+    rateLimit: (options?: fastifyRateLimit.RateLimitOptions) => preHandlerAsyncHookHandler;
   }
 }
 
-export interface FastifyRateLimitOptions {}
+type FastifyRateLimit = FastifyPluginCallback<fastifyRateLimit.RateLimitPluginOptions>;
 
-export interface errorResponseBuilderContext {
-  ban: boolean;
-  after: string;
-  max: number;
-  ttl: number;
-}
+declare namespace fastifyRateLimit {
 
-export interface FastifyRateLimitStoreCtor {
-  new (options: FastifyRateLimitOptions): FastifyRateLimitStore;
-}
+  export interface FastifyRateLimitOptions { }
 
-export interface FastifyRateLimitStore {
-  incr(
-    key: string,
-    callback: (
-      error: Error | null,
-      result?: { current: number; ttl: number }
-    ) => void
-  ): void;
-  child(
-    routeOptions: RouteOptions & { path: string; prefix: string }
-  ): FastifyRateLimitStore;
-}
+  export interface errorResponseBuilderContext {
+    ban: boolean;
+    after: string;
+    max: number;
+    ttl: number;
+  }
 
-interface DefaultAddHeaders {
-  'x-ratelimit-limit'?: boolean;
-  'x-ratelimit-remaining'?: boolean;
-  'x-ratelimit-reset'?: boolean;
-  'retry-after'?: boolean;
-}
+  export interface FastifyRateLimitStoreCtor {
+    new(options: FastifyRateLimitOptions): FastifyRateLimitStore;
+  }
 
-interface DraftSpecAddHeaders {
-  'ratelimit-limit'?: boolean;
-  'ratelimit-remaining'?: boolean;
-  'ratelimit-reset'?: boolean;
-  'retry-after'?: boolean;
-}
+  export interface FastifyRateLimitStore {
+    incr(
+      key: string,
+      callback: (
+        error: Error | null,
+        result?: { current: number; ttl: number }
+      ) => void
+    ): void;
+    child(
+      routeOptions: RouteOptions & { path: string; prefix: string }
+    ): FastifyRateLimitStore;
+  }
 
-interface DefaultAddHeadersOnExceeding {
-  'x-ratelimit-limit'?: boolean;
-  'x-ratelimit-remaining'?: boolean;
-  'x-ratelimit-reset'?: boolean;
-}
+  interface DefaultAddHeaders {
+    'x-ratelimit-limit'?: boolean;
+    'x-ratelimit-remaining'?: boolean;
+    'x-ratelimit-reset'?: boolean;
+    'retry-after'?: boolean;
+  }
 
-interface DraftSpecAddHeadersOnExceeding {
-  'ratelimit-limit'?: boolean;
-  'ratelimit-remaining'?: boolean;
-  'ratelimit-reset'?: boolean;
-}
+  interface DraftSpecAddHeaders {
+    'ratelimit-limit'?: boolean;
+    'ratelimit-remaining'?: boolean;
+    'ratelimit-reset'?: boolean;
+    'retry-after'?: boolean;
+  }
 
-export type RateLimitHook =
-  | 'onRequest'
-  | 'preParsing'
-  | 'preValidation'
-  | 'preHandler'
+  interface DefaultAddHeadersOnExceeding {
+    'x-ratelimit-limit'?: boolean;
+    'x-ratelimit-remaining'?: boolean;
+    'x-ratelimit-reset'?: boolean;
+  }
 
-export interface RateLimitOptions {
-  max?:
+  interface DraftSpecAddHeadersOnExceeding {
+    'ratelimit-limit'?: boolean;
+    'ratelimit-remaining'?: boolean;
+    'ratelimit-reset'?: boolean;
+  }
+
+  export type RateLimitHook =
+    | 'onRequest'
+    | 'preParsing'
+    | 'preValidation'
+    | 'preHandler'
+
+  export interface RateLimitOptions {
+    max?:
     | number
     | ((req: FastifyRequest, key: string) => number)
     | ((req: FastifyRequest, key: string) => Promise<number>);
-  timeWindow?: number | string;
-  hook?: RateLimitHook;
-  cache?: number;
-  store?: FastifyRateLimitStoreCtor;
-  /**
-   * @deprecated Use `allowList` property
-   */
-  whitelist?: string[] | ((req: FastifyRequest, key: string) => boolean);
-  allowList?: string[] | ((req: FastifyRequest, key: string) => boolean | Promise<boolean>);
-  continueExceeding?: boolean;
-  skipOnError?: boolean;
-  ban?: number;
-  onBanReach?: (req: FastifyRequest, key: string) => void;
-  keyGenerator?: (req: FastifyRequest) => string | number | Promise<string | number>;
-  errorResponseBuilder?: (
-    req: FastifyRequest,
-    context: errorResponseBuilderContext
-  ) => object;
-  enableDraftSpec?: boolean;
-  onExceeding?: (req: FastifyRequest, key: string) => void;
-  onExceeded?: (req: FastifyRequest, key: string) => void;
-}
+    timeWindow?: number | string;
+    hook?: RateLimitHook;
+    cache?: number;
+    store?: FastifyRateLimitStoreCtor;
+    /**
+     * @deprecated Use `allowList` property
+     */
+    whitelist?: string[] | ((req: FastifyRequest, key: string) => boolean);
+    allowList?: string[] | ((req: FastifyRequest, key: string) => boolean | Promise<boolean>);
+    continueExceeding?: boolean;
+    skipOnError?: boolean;
+    ban?: number;
+    onBanReach?: (req: FastifyRequest, key: string) => void;
+    keyGenerator?: (req: FastifyRequest) => string | number | Promise<string | number>;
+    errorResponseBuilder?: (
+      req: FastifyRequest,
+      context: errorResponseBuilderContext
+    ) => object;
+    enableDraftSpec?: boolean;
+    onExceeding?: (req: FastifyRequest, key: string) => void;
+    onExceeded?: (req: FastifyRequest, key: string) => void;
+  }
 
-export interface RateLimitPluginOptions extends RateLimitOptions {
-  global?: boolean;
-  cache?: number;
-  redis?: any;
-  nameSpace?: string;
-  addHeaders?: DefaultAddHeaders | DraftSpecAddHeaders;
-  addHeadersOnExceeding?:
+  export interface RateLimitPluginOptions extends RateLimitOptions {
+    global?: boolean;
+    cache?: number;
+    redis?: any;
+    nameSpace?: string;
+    addHeaders?: DefaultAddHeaders | DraftSpecAddHeaders;
+    addHeadersOnExceeding?:
     | DefaultAddHeadersOnExceeding
     | DraftSpecAddHeadersOnExceeding;
+  }
+  export const fastifyRateLimit: FastifyRateLimit
+  export { fastifyRateLimit as default }
 }
 
-declare const fastifyRateLimit: FastifyPluginCallback<RateLimitPluginOptions>;
-export default fastifyRateLimit;
+declare function fastifyRateLimit(...params: Parameters<FastifyRateLimit>): ReturnType<FastifyRateLimit>
+export = fastifyRateLimit
